+++ integration-tests/realm-apps/with-db/services/mongodb/config.json
@@ -8,7 +8,7 @@
       "partition": {
         "key": "partition",
         "type": "string",
-        "required": false,
+        "required": true,
         "permissions": {
           "read": true,
           "write": true
