+++ integration-tests/tests/src/tests/index.ts
@@ -56,5 +56,6 @@ import "./schema";
 import "./types";
 import "./sync/client-reset";
 import "./sync/open-behavior";
+import "./sync/realm-conversions";
 import "./sync/partition-values";
 import "./sync/sync-session";
