+++ integration-tests/tests/src/tests/sync/sync-session.ts
@@ -21,7 +21,8 @@ import { expect } from "chai";
 import { ConnectionState, ObjectSchema } from "realm";
 import { importAppBefore } from "../../hooks";
 import { DogSchema } from "../../schemas/person-and-dog-with-object-ids";
-import { generatePartition, randomVerifiableEmail } from "../../utils/generators";
+import { getRegisteredEmailPassCredentials } from "../../utils/credentials";
+import { generatePartition } from "../../utils/generators";
 import { sleep, throwAfterTimeout } from "../../utils/sleep";
 
 const DogForSyncSchema = {
@@ -31,7 +32,7 @@ const DogForSyncSchema = {
     _id: "objectId",
     breed: "string?",
     name: "string",
-    realm_id: "string?",
+    partition: { type: "string", default: "partition" },
   },
 };
 
@@ -44,22 +45,9 @@ const PersonForSyncSchema = {
     dogs: "Dog[]",
     firstName: "string",
     lastName: "string",
-    realm_id: "string?",
+    partition: { type: "string", default: "partition" },
   },
 };
-
-async function getRegisteredEmailPassCredentials(app: Realm.App) {
-  if (!app) {
-    throw new Error("No app supplied to 'getRegisteredEmailPassCredentials'");
-  }
-
-  const email = randomVerifiableEmail();
-  const password = "test1234567890";
-  // Create the user (see note in 'randomVerifiableEmail')
-  await app.emailPasswordAuth.registerUser({ email, password });
-
-  return Realm.Credentials.emailPassword(email, password);
-}
 
 function waitForConnectionState(session: Realm.App.Sync.Session, state: string) {
   return async () => {
@@ -76,7 +64,7 @@ function waitForConnectionState(session: Realm.App.Sync.Session, state: string) 
   };
 }
 
-function getSyncConfiguration(user: Realm.User, partition: any): Realm.Configuration {
+function getSyncConfiguration(user: Realm.User, partition: string): Realm.Configuration {
   const realmConfig = {
     schema: [
       {
@@ -86,7 +74,7 @@ function getSyncConfiguration(user: Realm.User, partition: any): Realm.Configura
           _id: "objectId",
           breed: "string?",
           name: "string",
-          realm_id: "string?",
+          partition: { type: "string", default: partition },
         },
       },
     ],
@@ -833,6 +821,7 @@ describe("SessionTest", () => {
             age: i,
             firstName: "John",
             lastName: "Smith",
+            partition: partition,
           });
         }
       });
@@ -1088,281 +1077,6 @@ describe("SessionTest", () => {
       }
 
       expect(realm3).to.be.undefined;
-    });
-  });
-
-  describe("Realm conversions", () => {
-    afterEach(() => Realm.clearTestState());
-    // TODO:  Realm.open() fails after export_to() of synced -> local
-    it.skip("valid conversions", async function (this: AppContext) {
-      const credentials = await getRegisteredEmailPassCredentials(this.app);
-      const credentials2 = await getRegisteredEmailPassCredentials(this.app);
-      const user = await this.app.logIn(credentials);
-
-      /*
-      Tests matrix:
-        1)   local, unencrypted Realm -> local, unencrypted Realm
-        2)   local, unencrypted Realm -> local, encrypted
-        3)   local, encrypted Realm -> local, encrypted Realm
-        4)   local, encrypted Realm -> local, unencrypted Realm
-
-        5)   local, unencrypted Realm -> synced, unencrypted Realm
-        6)   local, unencrypted Realm -> synced, encrypted
-        7)   local, encrypted Realm -> synced, encrypted Realm
-        8)   local, encrypted Realm -> synced, unencrypted Realm
-
-        9)   synced, unencrypted Realm -> local, unencrypted Realm
-        10)  synced, unencrypted Realm -> local, encrypted Realm
-        11)  synced, encrypted Realm -> local, unencrypted Realm
-        12)  synced, encrypted Realm -> local, encrypted Realm
-
-        13)  synced, unencrypted Realm -> synced, unencrypted Realm
-        14)  synced, unencrypted Realm -> synced, encrypted Realm
-        15)  synced, encrypted Realm -> synced, unencrypted Relam
-        16)  synced, encrypted Realm -> synced, encrypted Realm
-    */
-
-      const sourceLocation = ["local", "synced"];
-      const destinationLocation = ["local", "synced"];
-      const sourceEncryption = ["plain", "encrypted"];
-      const destinationEncryption = ["plain", "encrypted"];
-
-      const encryptionKey1 = new Int8Array(64);
-      for (let i = 0; i < 64; i++) {
-        encryptionKey1[i] = 1;
-      }
-      const encryptionKey2 = new Int8Array(64);
-      for (let i = 0; i < 64; i++) {
-        encryptionKey2[i] = 2;
-      }
-
-      // local realm that we will put a few dogs into
-      const configLocal: Realm.Configuration = {
-        schema: [PersonForSyncSchema, DogForSyncSchema],
-        path: "dogsLocal.realm",
-      };
-
-      // local, encrypted, realm that we will put a few dogs into
-      const configLocalEnc: Realm.Configuration = {
-        schema: [PersonForSyncSchema, DogForSyncSchema],
-        encryptionKey: encryptionKey1,
-        path: "dogsLocalEnc.realm",
-      };
-
-      // shim config for the synced realms
-      const configSync: Realm.Configuration = {
-        sync: {
-          // TODO: export_to doesn't fail is there's no user..?
-          user,
-          partitionValue: "foo",
-          //@ts-expect-error using internal method.
-          _sessionStopPolicy: "immediately", // Make it safe to delete files after realm.close()
-        },
-        schema: [DogForSyncSchema],
-        path: "dogsSynced.realm",
-      };
-
-      // shim config for the synced realms
-      const configSyncEnc: Realm.Configuration = {
-        sync: {
-          // TODO: export_to doesn't fail is there's no user..?
-          // TODO:  Realm.open() fails after export_to() of synced -> local
-          user,
-          partitionValue: "foo",
-          //@ts-expect-error using internal method.
-          _sessionStopPolicy: "immediately", // Make it safe to delete files after realm.close()
-        },
-        schema: [DogForSyncSchema],
-        encryptionKey: encryptionKey2,
-        path: "dogsSyncedEnc.realm",
-      };
-
-      // create five dogs in our local realm
-      const realmLocal = await Realm.open(configLocalEnc);
-      realmLocal.write(() => {
-        for (let i = 0; i < 5; i++) {
-          realmLocal.create("Dog", {
-            _id: new ObjectId(),
-            breed: "Domestic Short Hair",
-            name: `Brutus no. ${i}`,
-            realm_id: "foo",
-          });
-        }
-      });
-      realmLocal.close();
-
-      // create five dogs in our local realm
-      const realmLocalEnc = await Realm.open(configLocal);
-      realmLocalEnc.write(() => {
-        for (let i = 0; i < 5; i++) {
-          realmLocalEnc.create("Dog", {
-            _id: new ObjectId(),
-            breed: "Domestic Short Hair",
-            name: `..Encrypted.. no. ${i}`,
-            realm_id: "foo",
-          });
-        }
-      });
-      realmLocalEnc.close();
-
-      const configSyncedDogs = configSync;
-      configSyncedDogs.path = "dummy.realm";
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      configSyncedDogs.sync!.user = user;
-
-      // create five dogs in our synced realm
-      const realmSynced = await Realm.open(configSyncedDogs);
-      realmSynced.write(() => {
-        for (let i = 0; i < 5; i++) {
-          realmSynced.create("Dog", {
-            _id: new ObjectId(),
-            breed: "Fancy Long Hair",
-            name: `Darletta no. ${i}`,
-            realm_id: "foo",
-          });
-        }
-      });
-      realmSynced.close();
-
-      await user.logOut();
-
-      // FIXME:
-      // The test loops below should be refactored when moving to integration tests.
-      // Multiple tests should be generated instead of looping over the test combinations
-      // (see https://mochajs.org/#dynamically-generating-tests)
-      let testNo = 1;
-      for (const source of sourceLocation) {
-        for (const destination of destinationLocation) {
-          for (const srcEncryption of sourceEncryption) {
-            for (const dstEncryption of destinationEncryption) {
-              const configSrc = Object.assign(
-                {},
-                source == "local"
-                  ? srcEncryption == "plain"
-                    ? configLocal
-                    : configLocalEnc
-                  : srcEncryption == "plain"
-                  ? configSync
-                  : configSyncEnc,
-              );
-              const configDst = Object.assign(
-                {},
-                destination == "local"
-                  ? dstEncryption == "plain"
-                    ? configLocal
-                    : configLocalEnc
-                  : dstEncryption == "plain"
-                  ? configSync
-                  : configSyncEnc,
-              );
-
-              if (source == "synced") {
-                // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                configSrc.sync!.user = await this.app.logIn(credentials);
-              }
-              if (destination == "synced") {
-                // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                configDst.sync!.user = source == "synced" ? configSrc.sync!.user : await this.app.logIn(credentials2);
-              }
-
-              console.log(`\nTest ${testNo})  ${source}, ${srcEncryption}  -> ${destination}, ${dstEncryption}...\n`);
-
-              configDst.path += `_test_${testNo}.realm`;
-
-              if (srcEncryption == "encrypted") {
-                configSrc.encryptionKey = encryptionKey1;
-              }
-              if (dstEncryption == "encrypted") {
-                configDst.encryptionKey = encryptionKey2;
-              }
-
-              const realmSrc = await Realm.open(configSrc);
-              if (source == "synced") {
-                await realmSrc.syncSession?.uploadAllLocalChanges();
-                await realmSrc.syncSession?.downloadAllServerChanges();
-              }
-              realmSrc.writeCopyTo(configDst);
-              realmSrc.close();
-
-              const realmDst = await Realm.open(configDst);
-              if (destination == "synced") {
-                expect(realmDst.syncSession).to.not.be.undefined;
-                expect(realmDst.syncSession).to.not.be.null;
-                await realmDst.syncSession?.downloadAllServerChanges();
-                await realmDst.syncSession?.uploadAllLocalChanges();
-              }
-              realmDst.close();
-
-              if (source == "synced") {
-                await configSrc.sync?.user.logOut();
-              }
-              if (destination == "synced") {
-                await configDst.sync?.user.logOut();
-              }
-              Realm.deleteFile(configDst);
-
-              testNo++;
-            }
-          }
-        }
-      }
-    });
-
-    it("invalid conversions throw", async function (this: AppContext) {
-      // simple local realm
-      const configLocal = {
-        schema: [PersonForSyncSchema, DogForSyncSchema],
-        path: "dogsLocal.realm",
-      };
-
-      // user for flexible sync test
-      const credentials = await getRegisteredEmailPassCredentials(this.app);
-      const user = await this.app.logIn(credentials);
-
-      /*
-       *  Test 1:  check that `writeCopyTo` verifies parameter count and types
-       */
-      const realm = await Realm.open(configLocal);
-      expect(() => {
-        //@ts-expect-error too many arguments
-        realm.writeCopyTo("path", "encryptionKey", "invalidParameter");
-      }).throws("Invalid arguments: at most 1 expected, but 3 supplied.");
-      expect(() => {
-        //@ts-expect-error too few arguments
-        realm.writeCopyTo();
-      }).throws("Expected a config object");
-      expect(() => {
-        //@ts-expect-error wrong argument type
-        realm.writeCopyTo(null);
-      }).throws("`config` parameter must be an object");
-      expect(() => {
-        //missing `path` property
-        realm.writeCopyTo({});
-      }).throws("`path` property must exist in output configuration");
-      expect(() => {
-        //@ts-expect-error wrong `path` property type
-        realm.writeCopyTo({ path: 12345 });
-      }).throws("`path` property must be a string");
-      expect(() => {
-        //@ts-expect-error wrong `encryptionKey` property type
-        realm.writeCopyTo({ path: "outputPath", encryptionKey: "notBinary" });
-      }).throws("'encryptionKey' property must be an ArrayBuffer or ArrayBufferView");
-      expect(() => {
-        //@ts-expect-error wrong `sync` property type
-        realm.writeCopyTo({ path: "outputPath", sync: "invalidProperty" });
-      }).throws("'sync' property must be an object");
-      expect(() => {
-        realm.writeCopyTo({ path: "output", sync: { flexible: true, user } });
-      }).throws("Realm cannot be converted to a flexible sync realm unless flexible sync is already enabled");
-      /*
-       *  Test 2:  check that `writeCopyTo` can only be called at the right time
-       */
-      realm.write(() => {
-        expect(() => {
-          //@ts-expect-error wrong `sync` property type
-          realm.writeCopyTo({ path: "outputPath", sync: "invalidProperty" });
-        }).throws("Can only convert Realms outside a transaction");
-      });
     });
   });
 });
