+++ integration-tests/tests/src/utils/credentials.ts
@@ -1,0 +1,35 @@
+////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2023 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
+
+import { Credentials } from "realm";
+import { randomVerifiableEmail } from "./generators";
+
+export async function getRegisteredEmailPassCredentials(
+  app: Realm.App,
+): Promise<Credentials<Credentials.EmailPasswordPayload>> {
+  if (!app) {
+    throw new Error("No app supplied to 'getRegisteredEmailPassCredentials'");
+  }
+
+  const email = randomVerifiableEmail();
+  const password = "test1234567890";
+  // Create the user (see note in 'randomVerifiableEmail')
+  await app.emailPasswordAuth.registerUser({ email, password });
+
+  return Realm.Credentials.emailPassword(email, password);
+}
